backend/requirements.txt
django
djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt
djoser
pandas